OSD: block pg creation if there is a pending split on an ancestor
This is a hack to handle the (hopefully) rare case where we attempt
to recreate a pg which has been removed and then split.  We cannot
resurrect this pg, and we cannot create any pgs which would be
children since there would be hobject_t overlap.  A less bad
solution would be to delay the message which wants to create the pg,
but that would be a lot of machinery for a rare case.

Signed-off-by: Samuel Just <sam.just@inktank.com>